Committing changes done on the Authorization side of the application
Changes made to curl-scrips/auth files, added curl scrips to all.
Changes made to assets/scripts/auth- calling the api in api.js,
event handler callbacks added for auth processes in events.js,
handled response in ui.js for auth processes.
Changes made to assets/scripts/templates include to app.js in order to
add event listeners, in config.js I added the URL paths to our project
api to utilize in api.js.
In tic-tac-toe-client/index.html I added buttons and forms in order
to handle all of the events described above as well as forms and buttons
for some of the post-log-in activities that the user will be able to do.