Committing new progress with tic-tac-toe, new functions
On app.js I added an event listener for click. on api.js i added a new
game function which accesses the api and generates new game event, I
also created a gameupdate function which updates the array in the cells
with the selection of the user. in ui.js I updated the the messages
which will display when game is created and updated. I also added a
line which will store the current situation in the game inside of the
array.

DISREGARD NEW-GAME.SH edit.
DISREGARD LOGIC.JS FOR NOW.