Commiting additions to project
Added and fixed the api events for get new game, also added a function
in api to be able to process the information for getAllgames and to
display those games. In events.js i added information to display games
and also corrected some issues with my X's and O's. in ui.js I added the
handles for generating the games and also fixed for my events.js to fix
my logic a bit.