Added box update logic
Added logic to events.js to check if the game is over (winner was found)
if so, the game will update the user that there is already a winner.
Also added a check to see if the index is taken, then let the player
know that the space trying to fill is already occupied.